# SILOptimizer: try inplace_function for compile performance (Test)