I was profiling some slow compiles and saw that the std::functions in UpdatingListIterator seems to be incurring a lot of new/delete overhead (around 5% of the compile time for a -O -wmo build), so was curious as to how much moving the out-of-line std::function allocations inline would affect performance.

I haven't been able to test performance on code locally due to difficulties in building the SwiftPM part of a toolchain (although the Swift tests pass), and I don't expect this to be merged as-is (since I’m guessing the inplace_function dependency might be an issue). It'd be great if someone could trigger a CI compiler performance test to see if this is worth pursuing; otherwise I can try to fix the toolchain build errors and get some local numbers sometime later this week.

inplace_function.h was taken from https://github.com/WG21-SG14/SG14/blob/master/SG14/inplace_function.h and hackily modified to assert on exceptions.