@troughton 2022-04-08
I've run some tests locally and seem to be getting roughly a 2-4% speedup in the full build time on the project where I noticed the compile time issues. 
@eeckstein or @gottesmm (as the last people to modify these files): if you think it's worth running a full compiler performance test on CI then I'd appreciate you starting that; otherwise, given these are fairly modest speedups, I'm also okay with closing this PR.