@eeckstein 2022-04-11
nice work!
Eventually we want to get rid of InstModCallbacks at all. But in the meantime it sounds like a good compile time improvement.
@atrick any thoughts?