@atrick 2022-04-11
Thank you! Although this overhead is not surprising, it's unacceptable--we need a fix.

I really am surprised that std::function doesn't have any inline storage optimization though! llvm::unique_function does. Can you just use that instead?

I don't know if there's any possibility of including a header under a different software license. AFAIK we need to keep everything in the Swift repo under the same license. You'd need to check on the forum.

Just to make sure I understand, is the 5% overhead mostly from std::function allocation within UpdatingInstructionIterator here:

```
  UpdatingInstructionIterator.h
    callbacks(InstModCallbacks()
                    .onDelete([this](SILInstruction *toDelete) {
                      notifyDelete(toDelete);
                      toDelete->eraseFromParent();
                    })
                    .onCreateNewInst(
                        [this](SILInstruction *newlyCreatedInst) {
                          notifyNew(newlyCreatedInst);
                        }))
```

For the record, the InstructionDeleter could be fixed to directly call iteratorRegistry.notifyDelete and notifyNew. But that's unsafe within any SIL passes that still rely on InstModCallbacks. To do that safely, we would need to eliminate InstModCallbacks complete, which still requires fixing a handful of SIL passes.