@troughton 2022-04-12
I’ve looked into this a little more and now think that the timing improvements from before may have been benchmark noise. The ~5% impact was measured on the 5.6 branch and was mainly in `UpdatingInstructionIteratorRegistry::rechainCallbacks`; that particular compile time issue I was seeing seems to have been fixed by https://github.com/apple/swift/commit/d418192ceb777e1e99d4fc5cde77f82a0bc7442d. 

In terms of the timing from before: I ran the build a few more times with everything else closed on my machine, and the results were:

Without change:

```
swift build --configuration release  1081.76s user 34.91s system 579% cpu 3:12.78 total
swift build --configuration release  1070.14s user 34.85s system 570% cpu 3:13.81 total
swift build --configuration release  1071.76s user 33.99s system 572% cpu 3:13.10 total
```

With change:

```
swift build --configuration release  1065.35s user 34.08s system 565% cpu 3:14.58 total
swift build --configuration release  1066.52s user 33.85s system 568% cpu 3:13.43 total
swift build --configuration release  1067.52s user 35.01s system 568% cpu 3:14.03 total
```

That seems close enough to be in the margin of error, although I do find it interesting that the user time and CPU usage were consistently _slightly_ lower than without the changes in this PR. Either that's just benchmark noise or is a tiny real improvement caused by some other factors (alignment changes?), but either way I don't feel it's enough to justify this PR.

To quickly address one other point:

> I really am surprised that std::function doesn’t have any inline storage optimization though!

My understanding (from https://stackoverflow.com/a/46163732 and its comments) is `std::function` _does_ have an inline storage optimisation (up to a libstdc++/libc++ determined size), but it was (in 5.6) being exceeded in some places. 

Anyway, thank you for having fixed the underlying issue back in https://github.com/apple/swift/pull/41239 and I hope looking at this didn't take up much of your time.