@eeckstein 2022-04-12
Thanks for doing these experiments!