@atrick 2022-04-12
@eeckstein fixed a severe problem where these closures were allocating unnecessarily sometime in the 5.7 time frame. It should be in that file's git log. I suspect that's what you were running into.

Thanks for checking!